Location/backend/requirements.txt
Django==5.2.11
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.4.0
psycopg2-binary==2.9.9
python-decouple==3.8
Pillow==10.4.0
celery==5.4.0
redis==5.1.1
openpyxl==3.1.5
reportlab==4.2.0
pandas==2.2.3
drf-spectacular==0.27.0
django-auditlog==2.3.0
django-ipware==6.0.0
user-agents==2.2.0
django-extensions==3.2.3
minio==7.2.0
boto3==1.35.26
django-storages==1.14.6
gunicorn==21.2.0